+++ packages/browser-image-resizer-ex/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @browser-mc/browser-image-resizer-ex
 
+## 0.6.0
+
+### Minor Changes
+
+- Add `lanczos3` planar resizing and make it the default raw planar resize algorithm.
+
+### Patch Changes
+
+- Updated dependencies
+  - @browser-mc/webcodecs-color@0.2.0
+  - @browser-mc/webcodecs-avif@0.4.2
+
 ## 0.5.2
 
 ### Patch Changes