+++ packages/browser-movie-converter/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @browser-mc/browser-movie-converter
 
+## 0.2.0
+
+### Minor Changes
+
+- Add `lanczos3` planar resizing and make it the default raw planar resize algorithm.
+
+### Patch Changes
+
+- Updated dependencies
+  - @browser-mc/webcodecs-color@0.2.0
+
 ## 0.1.1
 
 ### Patch Changes